refactor(table): 주문 모듈 분리
