+++ order/src/main/java/kitchenpos/order/order/domain/OrderLineItems.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.order.domain;
 
 import java.util.Collections;
 import java.util.List;
