+++ order/src/main/java/kitchenpos/order/order/ui/request/OrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.ui.request;
+package kitchenpos.order.order.ui.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
