+++ order/src/main/java/kitchenpos/order/order/ui/request/OrderStatusRequest.java
@@ -1,9 +1,9 @@
-package kitchenpos.order.ui.request;
+package kitchenpos.order.order.ui.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.order.domain.OrderStatus;
 
 public class OrderStatusRequest {
