+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'kitchenpos'
-include 'table', 'common', 'menu'
+include 'table', 'common', 'menu', 'order'
